requirements.txt
solana>=0.30.0
solders>=0.18.0
anchorpy>=0.18.0
websockets>=11.0
aiohttp>=3.8.0
pytest-asyncio>=0.23.0
# jupiter-python-sdk>=1.0.0  # if/when available
